+++ src/lib/components/ModifyForm.svelte
@@ -13,8 +13,11 @@
 	import {
 		checkNoOverlappingRsvs,
 		checkSpaceAvailable,
-		validateBuddies
 	} from '$lib/validationUtils.js';
+
+	import {
+		validateBuddies
+	} from '$utils/validation'
 
 	import {
 		addMissingFields,
