+++ src/lib/validationUtils.js
@@ -241,21 +241,3 @@ export function checkNoOverlappingRsvs(settings, orig, sub, existing) {
 	}
 	return { status: 'success' };
 }
-
-export function validateBuddies(rsv) {
-	let userIds = Object.keys(get(users));
-	let validBuddies = [];
-	for (let buddy of rsv.buddies) {
-		if (rsv.user.id === buddy) {
-			return { status: 'error', msg: 'Cannot add yourself as a buddy' };
-		}
-		if (!userIds.includes(buddy)) {
-			return { status: 'error', msg: 'Unknown user in buddy field' };
-		}
-		if (validBuddies.includes(buddy)) {
-			return { status: 'error', msg: 'Duplicate buddies not allowed' };
-		}
-		validBuddies.push(buddy);
-	}
-	return { status: 'success' };
-}
