+++ tsconfig.json
@@ -13,7 +13,10 @@
 		"strict": true,
 		"types": ["vitest/globals", "@testing-library/jest-dom"],
 		"paths": {
+			"$lib":["./src/lib"],
 			"$lib/*": ["./src/lib/*"],
+			"$utils":["src/utils"],
+			"$utils/*": ["src/utils/*"],
 			"$types": ["./src/types.ts"]
 		},
 		"typeRoots": ["./node_modules/@types", "./types"]
